% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr_example}
\alias{expr_example}
\title{Example transformed RNA-Seq data from paired observations}
\format{
A list with 2 named objects:
\describe{
  \item{vst_expr}{VST transformed expression}
  \item{sample_meta_data}{data frame with variables for group, time, and subject ID.  Rows are in the same order as columns of vst_expr}
  ...
}
}
\usage{
expr_example
}
\description{
A list that has 2 components: A matrix with transformed RNA-Seq values
(simulated from a negative binomial and then transformed using the VST in the DESeq2 package) and
a data frame that has the sample meta data
}
\keyword{datasets}
