% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq-pkg.R
\docType{package}
\name{lmerSeq-pkg}
\alias{lmerSeq-pkg}
\alias{lmerSeq-pkg-package}
\title{lmerSeq: Wrappers for Fitting Linear Mixed Models to RNA-Seq Data}
\description{
Fitting utilities for estimating linear mixed models for use
with RNA-Seq data from studies with clustered/longitudinal study designs
}
