% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq_fit.R
\name{lmerSeq.fit}
\alias{lmerSeq.fit}
\title{Function to Fit linear mixed models to transformed RNA-Seq data}
\usage{
lmerSeq.fit(form = NULL, expr_mat = NULL, gene_names = NULL,
  sample_data = NULL, REML = T)
}
\arguments{
\item{form}{A one-sided linear formula describing both the fixed-effects and random-effects parts of the model using the syntax of the lme4 package}

\item{expr_mat}{A (G x N) numeric matrix or data frame of transformed RNA-seq counts (e.g. using VST from DESeq2), with genes in rows and samples in columns. G = number of genes.  N = number of samples.}

\item{gene_names}{An optional character vector of gene names (length G).  If unspecified, row names from the expression matrix will be used.}

\item{sample_data}{Data frame with N rows containing the fixed- and random-effects terms included in the formula.  The rows of the data frame must correspond (and be in the same order as) the columns of the expression matrix.}

\item{REML}{Should the models be fit with REML or regular ML?}
}
\description{
Wrapper function that its linear mixed models to (transformed) RNA-Seq data using the utilities present in the lme4 package.
}
