% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq_summary.R
\name{lmerSeq.summary}
\alias{lmerSeq.summary}
\title{Function to Fit MCMSeq Models}
\usage{
lmerSeq.summary(lmerSeq_results = NULL, coefficient = NULL,
  p_adj_method = "BH", ddf = "Satterthwaite", sort_results = T)
}
\arguments{
\item{lmerSeq_results}{Results object from running lmerSeq.fit}

\item{coefficient}{Character string or numeric indicator of which coefficient to summarize}

\item{p_adj_method}{Method for adjusting for multiple comparisons (default is Benjamini-Hochberg). See p.adjust.methods}

\item{ddf}{Method for computing degrees of freedom and t-statistics. Options are "Satterthwaite" and "Kenward-Roger"}

\item{sort_results}{Should the results table be sorted by adjusted p-value?}
}
\description{
Fits negative binomial generlized linear models and negative binomial generalized linear mixed models to RNA-Seq data using MCMC.
}
