% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq_contrast.R
\name{lmerSeq.contrast}
\alias{lmerSeq.contrast}
\title{Function to Fit MCMSeq Models}
\usage{
lmerSeq.contrast(lmerSeq_results = NULL, contrast_mat = NULL,
  p_adj_method = "BH", ddf = "Satterthwaite", sort_results = T)
}
\arguments{
\item{lmerSeq_results}{Results object from running lmerSeq.fit}

\item{contrast_mat}{Numeric matrix representing the contrast to be tested.  Matrix must have the same number of columns as the number of coefficients in the model.  If the matrix has multiple rows, a simultaneous F-test will be done}

\item{p_adj_method}{Method for adjusting for multiple comparisons (default is Benjamini-Hochberg). See p.adjust.methods}

\item{ddf}{Method for computing degrees of freedom and t-statistics. Options are "Satterthwaite" and "Kenward-Roger"}

\item{sort_results}{Should the results table be sorted by adjusted p-value?}
}
\description{
Fits negative binomial generlized linear models and negative binomial generalized linear mixed models to RNA-Seq data using MCMC.
}
