% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq_fit.R
\name{lmerSeq.fit}
\alias{lmerSeq.fit}
\title{Function to Fit linear mixed models to transformed RNA-Seq data}
\usage{
lmerSeq.fit(form = NULL, expr_mat = NULL, gene_names = NULL,
  sample_data = NULL, REML = T, parallel = F, cores = 2)
}
\arguments{
\item{form}{A one-sided linear formula describing both the fixed-effects and random-effects parts of the model using the syntax of the lme4 package}

\item{expr_mat}{A (G x N) numeric matrix or data frame of transformed RNA-seq counts (e.g. using VST from DESeq2), with genes in rows and samples in columns. G = number of genes.  N = number of samples.}

\item{gene_names}{An optional character vector of gene names (length G).  If unspecified, row names from the expression matrix will be used.}

\item{sample_data}{Data frame with N rows containing the fixed- and random-effects terms included in the formula.  The rows of the data frame must correspond (and be in the same order as) the columns of the expression matrix.}

\item{REML}{Should the models be fit with REML or regular ML?}

\item{parallel}{If on Mac or linux, use forking (via mclapply) to parallelize fits}

\item{cores}{Number of cores to use (default is 2)}
}
\description{
Wrapper function that its linear mixed models to (transformed) RNA-Seq data using the utilities present in the lme4 package.
}
\examples{
data("expr_data")
vst_expr <- expr_example$vst_expr
sample_meta_data <- expr_example$sample_meta_data

##  Only including 10 genes in the expression matrix
vst_expr <- vst_expr[1:10, ]

##  Fit the Model
fit.lmerSeq <- lmerSeq.fit(form = ~ group * time + (1|ids),
                           expr_mat = vst_expr,
                           sample_data = sample_meta_data,
                           REML = T)

}
