% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSeq_summary.R
\name{lmerSeq.summary}
\alias{lmerSeq.summary}
\title{Function to summarize individual regression coefficients}
\usage{
lmerSeq.summary(
  lmerSeq_results = NULL,
  coefficient = NULL,
  p_adj_method = "BH",
  ddf = "Satterthwaite",
  sort_results = T
)
}
\arguments{
\item{lmerSeq_results}{Results object from running lmerSeq.fit}

\item{coefficient}{Character string or numeric indicator of which coefficient to summarize}

\item{p_adj_method}{Method for adjusting for multiple comparisons (default is Benjamini-Hochberg). See p.adjust.methods}

\item{ddf}{Method for computing degrees of freedom and t-statistics. Options are "Satterthwaite" and "Kenward-Roger"}

\item{sort_results}{Should the results table be sorted by adjusted p-value?}
}
\description{
Conducts t-tests on individual regression coefficients from fits done using lmerSeq_fit function
}
\examples{
data("expr_data")
vst_expr <- expr_example$vst_expr
sample_meta_data <- expr_example$sample_meta_data

##  Only including 10 genes in the expression matrix
vst_expr <- vst_expr[1:10, ]

##  Fit the Model
fit.lmerSeq <- lmerSeq.fit(form = ~ group * time + (1|ids),
                           expr_mat = vst_expr,
                           sample_data = sample_meta_data,
                           REML = T)

##  Summarize the time coefficient
model_sum <- lmerSeq.summary(lmerSeq_results = fit.lmerSeq,
                             coefficient = 3,
                             p_adj_method = 'BH',
                             ddf = 'Satterthwaite',
                             sort_results = T)
print(model_sum)

}
